--- optd/src/optimizer/errors.rs
@@ -1,6 +1,0 @@
-use crate::memo::MemoError;
-
-#[derive(Debug)]
-pub enum OptimizeError {
-    MemoError(MemoError),
-}
